Correctly deal with Sends.
When inspecting insert events, only send the text
to the window if the last character of the insert
is a newline.  This changes behavior slightly:
if you paste multi-line text where the last line does
not end in a newline then it is not sent untill you
press enter.  I think that is OK, and otherwise we
cannot implement Send events that send from the
Snarf buffer.

Fixes issue 1.