Better handling of the entry address on text deletion and entry.
First, don't move the entry address unless all text is typed
before the entry point.

Second, don't redraw the prompt on multi-chracter removals.
If the prompt is hilighted and text is entered then acme
sends this as 2 events: delete, followed by an add. If the
prompt is redrawn during the delete then the addresses will
be mucked up for the subsequent add.

Fixes issue 7.