Remove newline test; Actually use MsgTooLong error
In the case of the newline test, there was a case that was later allowed
by code. It may be that the allowed newlines should not be default and
the test should get put back in.

In the case of the message too long error, by giving the message in the
error, the wrong error was being propogated forward.

I also updated the test case to regexp the error string for the part
that matters and removed a double failure for unexpected EOF errors.