On receiving ERR_NOTONCHANNEL, close the channel window.
It's possible to open a window for a channel that is not JOINed.
One way is to do TOPIC <channel> for a non-JOINed channel.
Before this change, it was not possible to close such a window,
because instead of getting a PART reply, the server sends an ERR_NOTONCHANNEL.
This CL closes the channel window when it gets an ERR_NOTONCHANNEL.

Fixes #29.