Refactor prompt tracking.
Instead of trying to track it numerically, track it using relative addresses and regular expressions. The prompt is always at the start of a line, and it's the only way that > can appear at the start of a line. So we can look back from $ for ^>. If we find it, it's the prompt.

Since we now have a more robust way to detect the prompt, we can re-add it when deleted too: if we don't find ^> in the buffer, then add one to the end. That is now our prompt.